refactor(heartbeat): move prompt support helpers into bounded modules
